refactor(session): extract SessionPermissionService (#4251 step 3)
- Move pickNumberedApprovalOption, parseNumberedApprovalOptions,
  normalizeApprovalLabel, resolveApprovalInput to permissions.ts
- Create SessionPermissionService wrapping PermissionRequestManager
- SessionManager delegates permission resolve/approve/reject to service
- Re-export resolveApprovalInput and PermissionDecision from session.ts
- 72 lines removed from session.ts (1511 → 1439)
- 487 tests pass (31 test files, 0 failures)
- No behavioral changes — pure refactor
